Improve scanning accuracy
