+++ lib/utils/diff.py
@@ -23,16 +23,22 @@ import re
 class DynamicContentParser:
     def __init__(self, content1, content2):
         self._differ = difflib.Differ()
-        self._static_patterns = self.remove_dynamic_patterns(
-            self._differ.compare(content1.split(), content2.split())
-        )
-        self.base_content = content1
+        self._is_static = content1 == content2
+        self._base_content = content1
+
+        if not self._is_static:
+            self._static_patterns = self.remove_dynamic_patterns(
+                self._differ.compare(content1.split(), content2.split())
+            )
 
     def remove_dynamic_patterns(self, patterns):
         return [pattern for pattern in patterns if not pattern.startswith(("-", "+"))]
 
     def compare_to(self, content):
-        diff = self._differ.compare(self.base_content.split(), content.split())
+        if self._is_static:
+            return content == self._base_content
+
+        diff = self._differ.compare(self._base_content.split(), content.split())
         return self._static_patterns == self.remove_dynamic_patterns(diff)
